+++ include/optionx_cpp/data/trading/TradeResultQuery.hpp
@@ -11,7 +11,7 @@ namespace optionx {
     /// \brief Identifies a broker-side trade whose final result should be fetched.
     class TradeResultQuery {
     public:
-        std::uint64_t trade_id = 0;   ///< Local persistent trade ID, used for tracing/result propagation.
+        std::uint32_t trade_id = 0;   ///< Local persistent trade ID, used for tracing/result propagation.
         std::int64_t option_id = 0;   ///< Numeric broker-side trade ID.
         std::string option_hash;      ///< String/hash broker-side trade ID.
         int retry_attempts = 15;      ///< Broker result retry attempts for temporarily empty/error responses.
